requirements.txt
matplotlib>=3.4.2
pandas>=1.2.4
plotly>=4.14.3
watchdog>=2.1.2
streamlit>=0.82.0
google-api-core>=1.25.1
google-cloud-storage>=1.38.0
google-auth>=1.24.0
google-cloud-bigquery>=2.7.0
google-cloud-core>=1.5.0
google-crc32c>=1.1.2
proto-plus>=1.13.0
six>=1.15.0
attrs>=21.2.0
python-pachyderm
python-decouple